+++ synapse-core/src/main/java/de/otto/synapse/annotation/EnableEventSource.java
@@ -80,4 +80,4 @@ public @interface EnableEventSource {
      */
     String messageLogReceiverEndpoint() default "";
 
-    }
+}