+++ synapse-edison/src/test/java/de/otto/synapse/edison/journal/JournalRestControllerTest.java
@@ -2,7 +2,7 @@ package de.otto.synapse.edison.journal;
 
 import de.otto.synapse.edison.state.StateRepositoryUiController;
 import de.otto.synapse.journal.Journal;
-import de.otto.synapse.journal.Journals;
+import de.otto.synapse.journal.JournalRegistry;
 import de.otto.synapse.message.TextMessage;
 import de.otto.synapse.messagestore.MessageStoreEntry;
 import org.junit.Before;
@@ -48,7 +48,7 @@ public class JournalRestControllerTest {
     private WebApplicationContext context;
 
     @MockBean
-    private Journals journals;
+    private JournalRegistry journals;
 
     private MockMvc mockMvc;
 
@@ -60,7 +60,7 @@ public class JournalRestControllerTest {
                 .thenReturn(Stream.of(
                         MessageStoreEntry.of("test", TextMessage.of("first", null)))
                 );
-        when(journals.containsKey("test")).thenReturn(true);
+        when(journals.hasJournal("test")).thenReturn(true);
         when(journals.getJournal("test")).thenReturn(Optional.of(journal));
     }