+++ synapse-examples/example-edison/src/main/java/de/otto/synapse/example/edison/state/EventConsumers.java
@@ -3,9 +3,7 @@ package de.otto.synapse.example.edison.state;
 import de.otto.synapse.annotation.EventSourceConsumer;
 import de.otto.synapse.example.edison.payload.BananaPayload;
 import de.otto.synapse.example.edison.payload.ProductPayload;
-import de.otto.synapse.journal.JournalingStateRepository;
 import de.otto.synapse.message.Message;
-import de.otto.synapse.messagestore.MessageStore;
 import de.otto.synapse.state.StateRepository;
 import org.slf4j.Logger;
 
@@ -14,13 +12,13 @@ import java.util.Optional;
 import static de.otto.synapse.example.edison.state.BananaProduct.bananaProductBuilder;
 import static org.slf4j.LoggerFactory.getLogger;
 
-public class BananaProductStateRepository extends JournalingStateRepository<BananaProduct> {
+public class EventConsumers {
 
-    private static final Logger LOG = getLogger(BananaProductStateRepository.class);
+    private static final Logger LOG = getLogger(EventConsumers.class);
+    private final StateRepository<BananaProduct> stateRepository;
 
-    public BananaProductStateRepository(final StateRepository<BananaProduct> delegate,
-                                        final MessageStore journalMessageStore) {
-        super(delegate, journalMessageStore);
+    public EventConsumers(final StateRepository<BananaProduct> bananaProductStateRepository) {
+        this.stateRepository = bananaProductStateRepository;
     }
 
     @EventSourceConsumer(
@@ -30,7 +28,7 @@ public class BananaProductStateRepository extends JournalingStateRepository<Bana
     public void consumeBananas(final Message<BananaPayload> message) {
         final String entityId = message.getKey().partitionKey();
 
-        final Optional<BananaProduct> computed = compute(entityId, (s, bananaProduct) -> {
+        final Optional<BananaProduct> computed = stateRepository.compute(entityId, (s, bananaProduct) -> {
             final BananaProduct.Builder builder = bananaProduct.isPresent()
                     ? bananaProductBuilder(bananaProduct.get())
                     : bananaProductBuilder();
@@ -49,7 +47,7 @@ public class BananaProductStateRepository extends JournalingStateRepository<Bana
     )
     public void consumeProducts(final Message<ProductPayload> message) {
         final String entityId = message.getKey().partitionKey();
-        final Optional<BananaProduct> computed = compute(entityId, (s, bananaProduct) -> {
+        final Optional<BananaProduct> computed = stateRepository.compute(entityId, (s, bananaProduct) -> {
             final BananaProduct.Builder builder = bananaProduct.isPresent()
                     ? bananaProductBuilder(bananaProduct.get())
                     : bananaProductBuilder();